SF-962 Store question audio for offline use (#696)
* SF-962 Store question audio for offline use

* cache question audio when navigating to checking app and online
* requesting question audio through audio service returns cached audio
* question audio in answer panel and question dialog works offline
* if audio url has changed the audio service will refresh its cache
* audio of answers remain unchanged

* Subscribe to cache question audio if offine for current project only

* Add additional comments for audio service

* Hook into the change detection cycle to refresh audio cache

* wip subscribe to all question doc edits

* Cache question audio when realtime query remoteDocChanged emits

* Correct blob prefix and remoteDocChanges observable

* Only update url in answers component if audio url is defined

+++ src/SIL.XForge.Scripture/ClientApp/src/app/core/sf-project.service.ts
@@ -11,6 +11,7 @@ import {
 } from 'realtime-server/lib/scriptureforge/models/sf-project-user-config';
 import { getTextDocId } from 'realtime-server/lib/scriptureforge/models/text-data';
 import { Canon } from 'realtime-server/lib/scriptureforge/scripture-utils/canon';
+import { AudioService } from 'xforge-common/audio.service';
 import { AuthService } from 'xforge-common/auth.service';
 import { CommandService } from 'xforge-common/command.service';
 import { AudioData } from 'xforge-common/models/audio-data';
@@ -40,39 +41,40 @@ export class SFProjectService extends ProjectService<SFProject, SFProjectDoc> {
     commandService: CommandService,
     private readonly pwaService: PwaService,
     private readonly authService: AuthService,
+    private readonly audioService: AudioService,
     http: HttpClient,
     private readonly machineHttp: MachineHttpClient
   ) {
     super(realtimeService, commandService, SF_PROJECT_ROLES, http);
+
     this.subscribe(this.pwaService.onlineStatus, async isOnline => {
       // Wait until logged in so that the remote store gets initialized
       if (isOnline && (await this.authService.isLoggedIn)) {
         const audioData = await this.realtimeService.offlineStore.getAllData<AudioData>(AudioData.COLLECTION);
         for (const audio of audioData) {
-          if (audio.deleteRef != null) {
+          if (audio.deleteRef != null && audio.projectRef != null) {
             await this.onlineDeleteAudio(audio.projectRef, audio.id, audio.deleteRef);
-            this.realtimeService.removeOfflineData(AudioData.COLLECTION, audio.id);
+            await this.realtimeService.removeOfflineData(AudioData.COLLECTION, audio.id);
             continue;
           }
-          const questionDoc = await this.realtimeService.subscribe<QuestionDoc>(
-            QuestionDoc.COLLECTION,
-            audio.realtimeDocRef!
-          );
-          if (questionDoc.data != null) {
-            const url = await this.onlineUploadAudio(
-              audio.projectRef,
-              audio.id,
-              new File([audio.blob!], audio.filename!)
-            );
-            if (questionDoc.data.dataId === audio.id) {
-              // The audio belongs to the question
-              questionDoc.submitJson0Op(op => op.set(qd => qd.audioUrl!, url));
-            } else {
-              const answerIndex = questionDoc.data.answers.findIndex(a => a.dataId === audio.id);
-              questionDoc.submitJson0Op(op => op.set(qd => qd.answers[answerIndex].audioUrl!, url));
+          if (audio.onlineUrl == null && audio.projectRef != null && audio.dataCollection === QuestionDoc.COLLECTION) {
+            // The audio file has not been uploaded to the server
+            const doc = this.realtimeService.get<QuestionDoc>(audio.dataCollection, audio.realtimeDocRef!);
+            let url: string;
+            if (doc.data != null) {
+              url = await this.onlineUploadAudio(audio.projectRef, audio.id, new File([audio.blob!], audio.filename!));
+              if (doc.data.dataId === audio.id) {
+                // The audio belongs to the question
+                doc.submitJson0Op(op => op.set(qd => qd.audioUrl!, url));
+                this.findOrUpdateAudioCache(audio.dataCollection, audio.id, url);
+              } else {
+                const answerIndex = doc.data.answers.findIndex(a => a.dataId === audio.id);
+                doc.submitJson0Op(op => op.set(qd => qd.answers[answerIndex].audioUrl!, url));
+                // Audio on answers are not cached
+                this.realtimeService.removeOfflineData(AudioData.COLLECTION, audio.id);
+              }
             }
           }
-          this.realtimeService.removeOfflineData(AudioData.COLLECTION, audio.id);
         }
       }
     });
@@ -159,6 +161,25 @@ export class SFProjectService extends ProjectService<SFProject, SFProjectDoc> {
     return (await this.onlineInvoke<string>('uninviteUser', { projectId, emailToUninvite }))!;
   }
 
+  /**
+   * Iterate through the questions in a project and cache the up-to-date audio data if it exists.
+   */
+  async onlineCacheAudio(questionDocs: Readonly<QuestionDoc[]>): Promise<void> {
+    for (const qd of questionDocs) {
+      if (qd.data != null) {
+        if (qd.data.isArchived) {
+          this.findOrUpdateAudioCache(qd.collection, qd.data.dataId, undefined);
+        } else {
+          this.findOrUpdateAudioCache(qd.collection, qd.data.dataId, qd.data.audioUrl);
+        }
+      }
+    }
+  }
+
+  findOrUpdateAudioCache(dataCollection: string, dataId: string, url?: string): Promise<AudioData | undefined> {
+    return this.audioService.findOrUpdateCache(dataCollection, dataId, url);
+  }
+
   async trainSelectedSegment(projectUserConfig: SFProjectUserConfig): Promise<void> {
     if (
       projectUserConfig.selectedTask !== 'translate' ||
@@ -227,13 +248,24 @@ export class SFProjectService extends ProjectService<SFProject, SFProjectDoc> {
    * Uploads the audio file to the file server, or if offline, stores the audio in IndexedDB and uploads
    * next time there is a valid connection.
    */
-  async uploadAudio(id: string, dataId: string, questionDocId: string, blob: Blob, filename: string): Promise<string> {
+  async uploadAudio(
+    id: string,
+    dataCollection: string,
+    dataId: string,
+    realtimeDocRef: string,
+    blob: Blob,
+    filename: string
+  ): Promise<string> {
     if (this.pwaService.isOnline) {
       // We are online. Upload directly to the server
-      return this.onlineUploadAudio(id, dataId, new File([blob], filename));
+      const onlineUrl = await this.onlineUploadAudio(id, dataId, new File([blob], filename));
+      if (dataCollection === QuestionDoc.COLLECTION && realtimeDocRef === getQuestionDocId(id, dataId)) {
+        await this.findOrUpdateAudioCache(dataCollection, dataId, onlineUrl);
+      }
+      return onlineUrl;
     }
     // Store the audio in indexedDB until we go online again
-    let localAudioData = AudioData.createUploadData(QuestionDoc.COLLECTION, dataId, id, questionDocId, blob, filename);
+    let localAudioData = AudioData.createUploadData(dataCollection, dataId, id, realtimeDocRef, blob, filename);
     localAudioData = await this.realtimeService.storeOfflineData(localAudioData);
     return URL.createObjectURL(localAudioData.blob);
   }
@@ -243,14 +275,17 @@ export class SFProjectService extends ProjectService<SFProject, SFProjectDoc> {
    * If the file is not present in IndexedDB, delete the audio file from the file server next time there is a
    * valid connection.
    */
-  async deleteAudio(id: string, dataId: string, ownerId: string): Promise<void> {
+  async deleteAudio(id: string, dataCollection: string, dataId: string, ownerId: string): Promise<void> {
     if (this.pwaService.isOnline) {
+      await this.findOrUpdateAudioCache(dataCollection, dataId, undefined);
       return this.onlineDeleteAudio(id, dataId, ownerId);
     }
-    // The audio existed locally and was never uploaded, remove it and return
-    if (await this.realtimeService.removeOfflineData(AudioData.COLLECTION, dataId)) {
+    const audio = await this.realtimeService.offlineStore.getData<AudioData>(AudioData.COLLECTION, dataId);
+    if (audio != null && audio.onlineUrl == null) {
+      // The audio existed locally and was never uploaded, remove it and return
+      await this.findOrUpdateAudioCache(audio.dataCollection, audio.id, undefined);
       return;
     }
-    this.realtimeService.storeOfflineData(AudioData.createDeletionData(QuestionDoc.COLLECTION, dataId, id, ownerId));
+    this.realtimeService.storeOfflineData(AudioData.createDeletionData(dataCollection, dataId, id, ownerId));
   }
 }
